Split out frontend routes to routes_root.go
